boot.scr: add support for dt overlays
Device-tree overlays make adjustments to dt's for add-in boards very
simple.

In order to support this you need to have a bootloader with
CONFIG_OF_LIBFDT_OVERLAY enabled to provide the 'fdt apply' command.

To make use of overlays the boot script will assumer fdt_overlays is a
list of dtb files to be applied as overlays.

Signed-off-by: Tim Harvey <tharvey@gateworks.com>