Makefile: change custom_rootfs hook to pass disk image
The eventual goal is to avoid requiring root (sudo) and loop devices to
build the root filesystem so lets head in that direction by passing only
the writable disk image to custom rootfs scripts which can then decide
to mount it or to use non-root tools such as e2tools to operate on
it.

Additionally add an error message and exit if a custom hook fails.

Signed-off-by: Tim Harvey <tharvey@gateworks.com>