Makefile: use e2tools to copy kernel files to pre-built ubuntu ext4 rootfs
Instead of using mount requiring root and available loop devices use a
prebuilt ubuntu ext filesystem and e2tools to copy the kernel files into
it.

To do this we start with a downloaded pre-built compressed ext filesystem.
Then we use e2freefrag to determine bytes available in the filesystem.
We next use 'du -bc' to determine the bytes needed for the files we are
going to add and bump that by 10% for margin. Finally we use e2fsk
followed by e2resize if needed then use e2mkdir and e2cp to copy the
files. Note that all files being added are owned by root.root.

It is not ideal to put multi-line shell script using temp variables
in Makefiles as it can be difficult to read/augment but these steps are
fairly simple. To avoid pitfalls:
 - escape scripts use of '$' with '$$'
 - convert the script to single-line by instering ';' between commands
   and escaping end-of-line with '\'
 - use $(expr <expression) instead of arithmetic statements ((...))
 - start with 'set -e' to ensure that a sub-command causes the make to
   fail

This also allows us to rid ourselves of pre-defining the filesystem size
as we determine the size necessary along the way resulting in a
minimally sized filesystem image. Note that the pre-built Gateworks
Ubuntu rootfs filesystems contain a script that runs on first boot
to resize the partition to the size of the device and the filesystem to
the size of the partition.

Signed-off-by: Tim Harvey <tharvey@gateworks.com>