Makefile: add kernel target to compare kernel version
Out of tree drivers depend on the kernel version they are built against
and often they only support a specific kernel version or range of
versions.

Add a kver_ge target to the Makefile that can evaluate the kernel
version in the linux directory and allow bracketing build steps based on
this.

Example:
ifeq ($(call kver_ge,6,6,0),0)
        @echo "kernel version ($(KMAJ),$(KMIN),$(KREV)) less than 6.6.0"
else
        @echo "kernel version ($(KMAJ),$(KMIN),$(KREV)) greather or equal than 6.6.0"
endif

Signed-off-by: Tim Harvey <tharvey@gateworks.com>